dockerfiles/fetcher/requirements.txt
python-dotenv
PyYAML==5.3.1
requests
